requirements.txt
# BewerbungsManager Web-App Dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
WTForms==3.1.1
Flask-WTF==1.2.1
email-validator==2.1.0
Jinja2==3.1.2
MarkupSafe==2.1.3
python-docx==1.1.0
reportlab==4.0.7
